Implement draft-acme-profiles-00
The ACME Profiles draft describes a way for ACME Servers to advertise
different certificate profiles that they're willing to issue. For
example, a Server might advertise a "classic" profile with 90-day
validity periods and OCSP URLs, and a "short-lived" profile with 6-day
validity and no OCSP URLs. The draft also describes how a client can
present these profiles to its operator and select a profile when making
a new-order request.

https://datatracker.ietf.org/doc/draft-aaron-acme-profiles/